document-api/requirements.txt
flask
transformers
torch
together
PyMuPDF
python-docx
python-dotenv

